+++ sound-mind-studio/tests/test_main_window.h
@@ -27,6 +27,10 @@ private slots:
     void exportTopmostLayerAudioNowFailsForAnUnrecognizedExtension();
     void exportTopmostLayerVideoNowFailsGracefullyWithNoContent();
     void exportTopmostLayerVideoNowExportsAnImportedLayer();
+    void exportTopmostLayerVideoAsyncRunsInTheBackgroundAndShowsTheCancelButton();
+    void exportTopmostLayerVideoAsyncCompletesSuccessfullyAndHidesTheCancelButton();
+    void cancelVideoExportStopsItAndDeletesThePartialFile();
+    void exportTopmostLayerVideoAsyncDoesNothingWhileAlreadyRunning();
     void importAudioFileShowsProgressThenCompletionInTheStatusBar();
     void poolTopmostLayerNowShowsProgressThenCompletionInTheStatusBar();
     void exportTopmostLayerAudioNowShowsProgressThenCompletionInTheStatusBar();